RISC-V: kprobes/kretprobe support
First implementation, adapted from arm64. The C.ADDISP16 instruction
gets simulated and the kprobes-handler called by inserting a C.EBREAK
instruction.

C.ADDISP16 was chosen as it sets-up the stack frame for functions.
Some work has been done to support probes on non-compressed
instructions but there is no support yet for decoding those.

The way forward should be to uncompress the instructions for simulation
to reduce the number of instructions used to decode the immediate
values on probe hit.

Signed-off-by: Patrick Stählin <me@packi.ch>